Move 'ignored' from qa_utils to misc_utils. Bump minor version for new functionality on branch.
