+++ test/test_qa_utils.py
@@ -1,5 +1,5 @@
 import re
-from dcicutils.qa_utils import mock_not_called, ignored
+from dcicutils.qa_utils import mock_not_called
 
 
 def test_mock_not_called():
@@ -12,12 +12,3 @@ def test_mock_not_called():
         assert m, "Expected assertion text did not appear."
     else:
         raise AssertionError("An AssertionError was not raised.")
-
-
-def test_ignored():
-
-    def foo(x, y):
-        ignored(x, y)
-
-    # Check that no error occurs for having used this.
-    assert foo(3, 4) is None
